# shared/api/bundles/order.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: OrderService API
  version: 1.0.0
  description: API управляет заказами на постройку космических кораблей.
tags:
  - name: Orders
    description: Операции управления заказами - создание, оплата, получение, отмена
paths:
  /api/v1/orders:
    post:
      summary: Создаёт новый заказ на основе выбранных пользователем деталей.
      operationId: CreateOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_order_request'
      responses:
        '200':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_order_response'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}/pay:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      summary: Проводит оплату ранее созданного заказа.
      operationId: OrderPay
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pay_order_request'
      responses:
        '200':
          description: Заказ успешно оплачен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pay_order_response'
        '400':
          description: Ошибка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    get:
      summary: Возвращает информацию о заказе.
      operationId: GetOrderByUuid
      tags:
        - Orders
      responses:
        '200':
          description: Информация о заказе полностью получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/get_order_response'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/orders/{order_uuid}/cancel:
    parameters:
      - $ref: '#/components/parameters/order_uuid'
    post:
      summary: Отменяет заказ
      operationId: OrderCancelByUuid
      tags:
        - Orders
      responses:
        '204':
          description: Заказ успешно отменён
        '400':
          description: Неизвестный статус заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Заказ уже оплачен и не может быть отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
components:
  schemas:
    create_order_request:
      type: object
      required:
        - user_uuid
        - part_uuids
      properties:
        user_uuid:
          type: string
          description: Uuid пользователя
        part_uuids:
          type: array
          items:
            type: string
          description: Uuids деталей
    generic_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 500
        message:
          type: string
          description: Описание ошибки
          example: An unexpected error occurred. Please try again later
    create_order_response:
      type: object
      required:
        - order_uuid
        - total_price
      properties:
        order_uuid:
          type: string
          description: Uuid заказа
        total_price:
          type: number
          format: float
          description: Общая стоимость деталей
          example: 123.45
    payment_method:
      type: string
      description: |
        Выбранный способ оплаты:
        * `UNKNOWN` — Неизвестный способ
        * `CARD` — Банковская карта
        * `SBP` — Система быстрых платежей
        * `CREDIT_CARD` — Кредитная карта
        * `INVESTOR_MONEY` — Деньги инвестора
      enum:
        - UNKNOWN
        - CARD
        - SBP
        - CREDIT_CARD
        - INVESTOR_MONEY
    pay_order_request:
      type: object
      required:
        - payment_method
      properties:
        payment_method:
          $ref: '#/components/schemas/payment_method'
    pay_order_response:
      type: object
      required:
        - transaction_uuid
      properties:
        transaction_uuid:
          type: string
    bad_request_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 400
        message:
          type: string
          description: Описание ошибки
          example: The request body is invalid or malformed
    not_found_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
          example: Order by uuid 'unknown' not found
    order_status:
      type: string
      description: |
        Статус заказа:
        * `PENDING_PAYMENT` - Ожидает оплату
        * `PAID` - Оплачен
        * `CANCELLED` - Отменен
      enum:
        - UNKNOWN
        - PENDING_PAYMENT
        - PAID
        - CANCELLED
    order_dto:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string
          description: Uuid заказа
        user_uuid:
          type: string
          description: Uuid пользователя
        part_uuids:
          type: array
          items:
            type: string
          description: Uuids деталей заказа
        total_price:
          type: number
          format: float
          description: Общая цена
          example: 123.45
        transaction_uuid:
          type: string
          description: Uuid транзакции
        payment_method:
          $ref: '#/components/schemas/payment_method'
        status:
          $ref: '#/components/schemas/order_status'
    get_order_response:
      type: object
      required:
        - order
      properties:
        order:
          $ref: '#/components/schemas/order_dto'
    conflict_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 409
        message:
          type: string
          description: Описание ошибки
          example: The order has already been paid for and cannot be cancelled
  parameters:
    order_uuid:
      name: order_uuid
      in: path
      required: true
      description: Uuid заказа для которого проводится оплата
      schema:
        type: string
x-ogen:
  target: ./shared/pkg/openapi/order/v1
  package: order_v1
  clean: true